+++ demo/src/app/components/+progress/index.ts
@@ -1,0 +1,4 @@
+import { ProgressBasicDemoComponent } from './basic/progress-basic-demo.component';
+import { ProgressStackedDemoComponent } from './stacked/progress-stacked-demo.component';
+
+export const PROGRESS_COMPONENTS = [ProgressBasicDemoComponent, ProgressStackedDemoComponent];