+++ demo/src/app/components/+progress/stacked/progress-stacked-demo.component.ts
@@ -1,0 +1,7 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'app-progress-stacked-demo',
+    templateUrl: './progress-stacked-demo.component.html'
+})
+export class ProgressStackedDemoComponent {}