+++ src/directive/index.ts
@@ -6,3 +6,4 @@ export * from './thy-enter.directive';
 export * from './thy-row.directive';
 export * from './thy-show';
 export * from './thy-stop-propagation.directive';
+export * from './thy-scroll.directive';