+++ src/editor/index.ts
@@ -1,4 +1,0 @@
-export * from './module';
-export * from './editor.component';
-export * from './editor.service';
-export * from './editor.constant';