+++ src/progress/interfaces.ts
@@ -1,0 +1,8 @@
+export type ThyProgressTypes = 'primary' | 'success' | 'info' | 'warning' | 'danger';
+
+export interface ThyStackedValue {
+    value: number;
+    type?: string;
+    color?: string;
+    label?: string;
+}