+++ src/select/index.ts
@@ -3,5 +3,5 @@ export * from './module';
 export * from './option.component';
 export * from './option-group.component';
 // export * from './select-container.component';
-export { ThySelectCustomComponent } from './select-custom.component';
+export { ThySelectCustomComponent } from './custom-select.component';
 export { ThySelectComponent } from './select.component';